requirements.txt
face-recognition==1.3.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy>=1.24.0,<2.0  # 必须使用1.x版本，dlib不兼容2.x
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
requests==2.31.0